+++ data/src/main/java/dk/itu/data/utils/RoutingUtils.java
@@ -9,7 +9,7 @@ public class RoutingUtils {
     private static final float EARTH_RADIUS_KM = 6371f;
 
     // This is based on the haversine formula to calculate distance between two points while accounting for earths curvature
-    public static Float distanceMeters(float lat1, float lon1, float lat2, float lon2){
+    public static Double distanceMeters(double lat1, double lon1, double lat2, double lon2){
         double lat1Rad = Math.toRadians(lat1);
         double lon1Rad = Math.toRadians(lon1);
         double lat2Rad = Math.toRadians(lat2);
@@ -22,7 +22,7 @@ public class RoutingUtils {
 
         double c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
 
-        return (float) (EARTH_RADIUS_KM * c * 1000);
+        return (EARTH_RADIUS_KM * c * 1000);
     }
 
     public static float distanceMetersFloat(double lat1, double lon1, double lat2, double lon2){
